put all login Test in one test lass. remove ChangepassTest class and ForgotPassTest
