+++ LV587SETOPENCART/LV587SETOPENCART/Tests/LoginTest.cs
@@ -11,7 +11,7 @@ namespace LV587SETOPENCART.Tests
 {
     [TestFixture]
     [AllureNUnit]
-    [AllureSuite("LoginTest")]
+    [AllureSuite("[LoginForm]")]
     [AllureDisplayIgnored]
     class LoginTest
     {
@@ -39,7 +39,7 @@ namespace LV587SETOPENCART.Tests
             classWithDriver.NavigateToURL();
         }
 
-        [Test]
+           [Test]
         public void LoginPageTest()
         {
             //Click on My Account > Login
@@ -52,6 +52,50 @@ namespace LV587SETOPENCART.Tests
             MyAccountPage myAccountPage = new MyAccountPage(driver);
             string expRes = "My Account";
             var actRes = myAccountPage.MyAccountText();
+            Assert.AreEqual(expRes, actRes);
+
+            headerComponent.ClickOnMyAccount(MyAccountMenuActions.Logout);
+        }
+
+
+        [Test]
+        public void ChangePassword()
+        {
+            //Click on My Account > Login
+            HeaderComponent headerComponent = new HeaderComponent(driver);
+            headerComponent.ClickOnMyAccount(MyAccountMenuActions.Login);
+            //login 
+            LoginBL loginBL = new LoginBL(driver);
+            loginBL.Login("user1@gmail.com", "qwertyasdf12345678");
+            //click password button on right side bar
+            RightSideBar rightSideBar = new RightSideBar(driver);
+            rightSideBar.PasswordListButtonClick();
+            //input new pass
+            ChangePassword changePassword = new ChangePassword(driver);
+            changePassword.InputChangePasswordText("qwertyasdf12345678");
+            changePassword.ClickContinueButtonChangePassword();
+            //Assert
+            string expRes = "Success: Your password has been successfully updated.";
+            var actRea = changePassword.AlertMessageText();
+            Assert.AreEqual(expRes, actRea);
+
+            headerComponent.ClickOnMyAccount(MyAccountMenuActions.Logout);
+        }
+
+
+        [Test]
+        public void ForgottenPasswordPageTest()
+        {
+            //Click on My Account > Login
+            HeaderComponent headerComponent = new HeaderComponent(driver);
+            headerComponent.ClickOnMyAccount(MyAccountMenuActions.Login);
+            // Click on "Forgotten Password" link text and input email
+            ForgottenPasswordBL forgottenPasswordBL = new ForgottenPasswordBL(driver);
+            forgottenPasswordBL.ForgottenPassword("user1@gmail.com");
+            //Assert
+            ForgottenPasswordPage forgottenPasswordPage = new ForgottenPasswordPage(driver);
+            string expRes = "An email with a confirmation link has been sent your email address.";
+            var actRes = forgottenPasswordPage.AlertMessageText();
             Assert.AreEqual(expRes, actRes);
         }
     }
